move classes from tools/pair_analysis into aisynphys/pair_analysis
